requirements.txt
numpy>=1.21
pandas>=1.5
pyyaml>=6.0
tqdm>=4.66
fastparquet
pyarrow
scipy>=1.10
scikit-learn>=1.2
statsmodels>=0.13
